fixed import file paths, renamed to index in each component file, pulled public and src into top level for now so app can run will need to get them back in client eventually
